Refactor remy.gui.filebrowser into remy.gui.browser.* modules
